example/requirements.txt
Flask-Authlib==1.2.0